pyproject.toml
[project]
name = "bil_hh_agent"
version = "0.1.0"
description = "ИИ-агент для поиска вакансий на hh.ru"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.26.1",
    "sqlalchemy>=2.0.20",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.6.0",
    "python-jose>=3.3.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "asyncpg>=0.27.0",
    "pyjwt>=2.10.1,<3.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "isort>=6.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "autoflake>=2.1.0",
    "bcrypt>=4.0.0"
]

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = false
force_grid_wrap = 0
use_parentheses = false
line_length = 120

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503"]

[tool.mypy]
python_version = 3.11
strict = true
ignore_missing_imports = true

[tool.poetry]
package-mode = false
